// models/enums.ts
export enum ExerciseType {
  Bodyweight = 'Bodyweight',
  Machine = 'Machine',
  FreeWeight = 'FreeWeight',
  Cardio = 'Cardio',
  Flexibility = 'Flexibility',
  Balance = 'Balance',
  Mobility = 'Mobility',
  Endurance = 'Endurance',
  Other = 'Other',
}

export enum ExerciseStyle {
  Isolation = 'Isolation',
  Compound = 'Compound',
  Accessory = 'Accessory',
  Olympic = 'Olympic',
  Bodybuilding = 'Bodybuilding',
  Powerlifting = 'Powerlifting',
  Strongman = 'Strongman',
  Other = 'Other',
}

export enum ExerciseMuscleGroup {
  UpperBody = 'UpperBody',
  LowerBody = 'LowerBody',
  FullBody = 'FullBody',
  Chest = 'Chest',
  Back = 'Back',
  Arms = 'Arms',
  Legs = 'Legs',
  Core = 'Core',
  Shoulders = 'Shoulders',
  Hips = 'Hips',
  Hamstrings = 'Hamstrings',
  Calves = 'Calves',
  Adductors = 'Adductors',
  Abductors = 'Abductors',
  Triceps = 'Triceps',
  Biceps = 'Biceps',
  Forearms = 'Forearms',
  Neck = 'Neck',
  UpperLat = 'UpperLat',
  LowerLat = 'LowerLat',
  Traps = 'Traps',
  MidBack = 'MidBack',
  Brachioradialis = 'Brachioradialis',
  LongHeadDominant = 'LongHeadDominant',
  LateralHeadDominant = 'LateralHeadDominant',
  RotatorCuff = 'RotatorCuff',
  Glutes = 'Glutes',
  Abs = 'Abs',
}

export enum ExerciseDominance {
  BicepsLateralHead = 'Biceps Lateral Head',
  BicepsLongHead = 'Biceps Long Head',
  Brachioradialis = 'Brachioradialis',
  TricepsLateralHead = 'Triceps Lateral Head',
  TricepsLongHead = 'Triceps Long Head',
  TricepsMedialHead = 'Triceps Medial Head',
  Quadriceps = 'Quadriceps',
  Glutes = 'Glutes',
  UpperChest = 'Upper Chest',
  MiddleChest = 'Middle Chest',
  LowerChest = 'Lower Chest',
  AnteriorDeltoid = 'AnteriorDeltoid',
  PosteriorDeltoid = 'PosteriorDeltoid',
  LateralDeltoid = 'LateralDeltoid',
}

export enum ExerciseMachine {
  Machine = 'Machine',
  Bodyweight = 'Bodyweight',
  SmithMachine = 'SmithMachine',
  Barbell = 'Barbell',
  Dumbbell = 'Dumbbell',
  TrapBar = 'TrapBar',
  Cable = 'Cable',
  Other = 'Other',
}

export enum HandPosition {
  Neutral = 'Neutral',
  Close = 'Close',
  Wide = 'Wide',
  Overhead = 'Overhead',
  Underhand = 'Underhand',
}

export enum ExerciseEquipment {
  Belt = 'Belt',
  KneeSleeves = 'Knee Sleeves',
  ElbowSleeves = 'Elbow Sleeves',
  WristSleeves = 'Wrist Sleeves',
  AnkleStraps = 'Ankle Straps',
  Straps = 'Straps',
  Bands = 'Bands',
}

export enum ExerciseHandle {
  ShortStraightBar = 'Short straight bar',
  LongStraightBar = 'Long straight bar',
  SingleRope = 'Single rope',
  DoubleRope = 'Double rope',
  DHandle = 'D handle',
  EzBar = 'Ez bar',
  TBarHandle = 'T-bar handle',
  R0T8Handle = 'R0-T8 handle',
  R0T8LongBar = 'R0-T8 long bar',
  R0T8ShortBar = 'R0-T8 short bar',
  R0T8SpreaderBar = 'R0-T8 spreader bar',
  R0T8DLandmineHandle = 'R0-T8 D landmine handle',
  KAZ = 'KAZ',
  Other = 'Other',
}

export enum ExerciseTier {
  S = 'S',
  A = 'A',
  B = 'B',
  C = 'C',
  F = 'F',
}

export enum RIR {
  One = '1',
  Two = '2',
  Three = '3',
  Four = '4',
  Five = '5',
  Six = '6',
  Seven = '7',
  Eight = '8',
  Nine = '9',
  Ten = '10',
  F = 'F',
  FF = 'FF',
}